--- a/includes/acl/compression/impl/quantize_streams.h
+++ b/includes/acl/compression/impl/quantize_streams.h
@@ -100,6 +100,12 @@ namespace acl
 
 			const transform_streams* raw_bone_streams;
 
+#ifdef ACL_COMPRESSION_OPTIMIZED
+
+			float* adjusted_shell_distances;		// 1 per transform per sample in segment
+
+#endif
+
 			rtm::qvvf* additive_local_pose;			// 1 per transform
 			rtm::qvvf* raw_local_pose;				// 1 per transform
 			rtm::qvvf* lossy_local_pose;			// 1 per transform
@@ -175,6 +181,12 @@ namespace acl
 				self_transform_indices = allocate_type_array<uint32_t>(allocator, num_bones);
 				chain_bone_indices = allocate_type_array<uint32_t>(allocator, num_bones);
 
+#ifdef ACL_COMPRESSION_OPTIMIZED
+
+				adjusted_shell_distances = allocate_type_array<float>(allocator, num_bones * clip_.segments->num_samples);
+
+#endif
+
 				for (uint32_t transform_index = 0; transform_index < num_bones; ++transform_index)
 				{
 					const transform_metadata& metadata_ = clip_.metadata[transform_index];
@@ -200,10 +212,28 @@ namespace acl
 				deallocate_type_array(allocator, parent_transform_indices, num_bones);
 				deallocate_type_array(allocator, self_transform_indices, num_bones);
 				deallocate_type_array(allocator, chain_bone_indices, num_bones);
+
+#ifdef ACL_COMPRESSION_OPTIMIZED
+
+				deallocate_type_array(allocator, adjusted_shell_distances, num_bones * clip.segments->num_samples);
+				
+#endif
+
 			}
 
 			void set_segment(segment_context& segment_)
 			{
+
+#ifdef ACL_COMPRESSION_OPTIMIZED
+
+				const uint32_t num_frames = segment_.num_samples;
+				ACL_ASSERT(num_frames <= clip.segments->num_samples, "Segment has too many samples!");
+				const auto calculate_object_space_distance_impl = std::mem_fn(has_scale ? &itransform_error_metric::calculate_object_space_distance : &itransform_error_metric::calculate_object_space_distance_no_scale);
+				itransform_error_metric::calculate_object_space_distance_args calculate_object_space_distance_args;
+				const clip_context::transform_link* transform_links = clip.transform_links;
+
+#endif
+
 				segment = &segment_;
 				bone_streams = segment_.bone_streams;
 				num_samples = segment_.num_samples;
@@ -278,6 +308,33 @@ namespace acl
 
 					uint8_t* sample_raw_object_transforms = raw_object_transforms + (sample_index * sample_transform_size);
 					local_to_object_space_impl(error_metric_, local_to_object_space_args_raw, sample_raw_object_transforms);
+
+#ifdef ACL_COMPRESSION_OPTIMIZED
+
+					float* adjusted_shell_distances_ = adjusted_shell_distances + sample_index;
+					for (uint32_t transform_index = 0; transform_index < num_bones; ++transform_index)
+					{
+						// Initialize adjusted_shell_distance.
+						adjusted_shell_distances_[transform_index * num_frames] = metadata[transform_index].shell_distance;
+					}
+					for (uint32_t transform_index = 0; transform_index < num_bones; ++transform_index)
+					{
+						// Propagate adjusted_shell_distance into ancestors.
+						const clip_context::transform_link& cur_link = transform_links[transform_index];
+						const uint32_t cur_parent = cur_link.parent_transform_index;
+						if (cur_parent != k_invalid_track_index)
+						{
+							const uint32_t cur_child = cur_link.child_transform_index;
+							calculate_object_space_distance_args.transform0 = sample_raw_object_transforms + cur_parent * metric_transform_size;
+							calculate_object_space_distance_args.transform1 = sample_raw_object_transforms + cur_child * metric_transform_size;
+							const float link_distance = rtm::scalar_cast(calculate_object_space_distance_impl(error_metric_, calculate_object_space_distance_args));
+							float& adjusted_parent_shell_distance = adjusted_shell_distances_[cur_parent * num_frames];
+							adjusted_parent_shell_distance = std::max(adjusted_parent_shell_distance, link_distance + adjusted_shell_distances_[cur_child * num_frames]);
+						}
+					}
+
+#endif
+
 				}
 			}
 
@@ -628,7 +685,13 @@ namespace acl
 			const itransform_error_metric* error_metric = context.error_metric;
 			const bool needs_conversion = context.needs_conversion;
 			const bool has_additive_base = context.has_additive_base;
+
+#ifndef ACL_COMPRESSION_OPTIMIZED
+
 			const transform_metadata& target_bone = context.metadata[target_bone_index];
+
+#endif
+
 			const uint32_t num_transforms = context.num_bones;
 			const size_t sample_transform_size = context.metric_transform_size * context.num_bones;
 			const float sample_rate = context.sample_rate;
@@ -655,7 +718,16 @@ namespace acl
 			itransform_error_metric::calculate_error_args calculate_error_args;
 			calculate_error_args.transform0 = nullptr;
 			calculate_error_args.transform1 = needs_conversion ? (const void*)(context.local_transforms_converted + (context.metric_transform_size * target_bone_index)) : (const void*)(context.lossy_local_pose + target_bone_index);
+
+#ifdef ACL_COMPRESSION_OPTIMIZED
+
+			const float* adjusted_shell_distances = context.adjusted_shell_distances + (target_bone_index * context.num_samples);
+
+#else
+
 			calculate_error_args.construct_sphere_shell(target_bone.shell_distance);
+
+#endif
 
 			const uint8_t* raw_transform = context.raw_local_transforms + (target_bone_index * context.metric_transform_size);
 			const uint8_t* base_transforms = context.base_local_transforms;
@@ -684,6 +756,12 @@ namespace acl
 					apply_additive_to_base_impl(error_metric, apply_additive_to_base_args_lossy, context.lossy_local_pose);
 				}
 
+#ifdef ACL_COMPRESSION_OPTIMIZED
+
+				calculate_error_args.construct_sphere_shell(adjusted_shell_distances[sample_index]);
+
+#endif
+
 				calculate_error_args.transform0 = raw_transform;
 				raw_transform += sample_transform_size;
 
@@ -710,7 +788,13 @@ namespace acl
 			const itransform_error_metric* error_metric = context.error_metric;
 			const bool needs_conversion = context.needs_conversion;
 			const bool has_additive_base = context.has_additive_base;
+
+#ifndef ACL_COMPRESSION_OPTIMIZED
+
 			const transform_metadata& target_bone = context.metadata[target_bone_index];
+
+#endif
+
 			const size_t sample_transform_size = context.metric_transform_size * context.num_bones;
 			const float sample_rate = context.sample_rate;
 			const float clip_duration = context.clip_duration;
@@ -744,7 +828,16 @@ namespace acl
 			itransform_error_metric::calculate_error_args calculate_error_args;
 			calculate_error_args.transform0 = nullptr;
 			calculate_error_args.transform1 = context.lossy_object_pose + (target_bone_index * context.metric_transform_size);
+
+#ifdef ACL_COMPRESSION_OPTIMIZED
+
+			const float* adjusted_shell_distances = context.adjusted_shell_distances + (target_bone_index * context.num_samples);
+
+#else
+
 			calculate_error_args.construct_sphere_shell(target_bone.shell_distance);
+
+#endif
 
 			const uint8_t* raw_transform = context.raw_object_transforms + (target_bone_index * context.metric_transform_size);
 			const uint8_t* base_transforms = context.base_local_transforms;
@@ -774,6 +867,12 @@ namespace acl
 				}
 
 				local_to_object_space_impl(error_metric, local_to_object_space_args_lossy, context.lossy_object_pose);
+
+#ifdef ACL_COMPRESSION_OPTIMIZED
+
+				calculate_error_args.construct_sphere_shell(adjusted_shell_distances[sample_index]);
+
+#endif
 
 				calculate_error_args.transform0 = raw_transform;
 				raw_transform += sample_transform_size;
@@ -1677,14 +1776,35 @@ namespace acl
 
 						// Calculate our error
 						const uint8_t* raw_frame_transform = raw_transform + (interp_frame_index * sample_transform_size);
+
+#ifdef ACL_COMPRESSION_OPTIMIZED
+
+						const float* adjusted_shell_distances = context.adjusted_shell_distances + interp_frame_index;
+
+#endif
+
 						for (uint32_t bone_index = 0; bone_index < num_bones; ++bone_index)
 						{
+
+#ifndef ACL_COMPRESSION_OPTIMIZED
+
 							const transform_metadata& target_bone = context.metadata[bone_index];
+
+#endif
 
 							itransform_error_metric::calculate_error_args calculate_error_args;
 							calculate_error_args.transform0 = raw_frame_transform + (bone_index * context.metric_transform_size);
 							calculate_error_args.transform1 = context.lossy_object_pose + (bone_index * context.metric_transform_size);
+
+#ifdef ACL_COMPRESSION_OPTIMIZED
+
+							calculate_error_args.construct_sphere_shell(adjusted_shell_distances[bone_index * num_frames]);
+
+#else
+
 							calculate_error_args.construct_sphere_shell(target_bone.shell_distance);
+
+#endif
 
 #if defined(RTM_COMPILER_MSVC) && defined(RTM_ARCH_X86) && RTM_COMPILER_MSVC == RTM_COMPILER_MSVC_2015
 							// VS2015 fails to generate the right x86 assembly, branch instead