https://github.com/nfrechette/acl/issues/353
https://github.com/nfrechette/acl/issues/373


**ACL_SJSON_FIX**

Fix that supports testing of sjson_file_type::raw_track_list, instead of only sjson_file_type::raw_clip.  

I needed this to test my edge case animation within ACL.


**ACL_COMPRESSION_OPTIMIZED**

Prevent low-magnitude channels from becoming constant when worst-case shell distance and object space distance are exceeded. 
Apply error correction after constant and default tracks are processed.
Propagate shell distance and object space distance through ancestors before compression.

Massive memory and quality upgrade.
In our use case, there's a **28%** reduction in memory overall(**33%** in humans, **22%** in creatures).
Our worst edge case used to compress in 30 seconds, and failed regression with an error of nearly 2 meters.  Now regression passes well within precision settings, and it **compresses 10x faster**.
Compression is faster generally, as seen in _py make.py -regression_test_:
```
uniformly_sampled_database.config.sjson: 04.04s -> 03.38s
uniformly_sampled_database_4kb.config.sjson: 04.03s -> 03.33s
uniformly_sampled_database_4kb_mixed.config.sjson: 04.73s  -> 04.04s
uniformly_sampled_database_mixed.config.sjson: 04.64s -> 03.99s
uniformly_sampled_mixed_var_0.config.sjson: 01.14s -> 01.26s
uniformly_sampled_mixed_var_1.config.sjson: 01.61s -> 01.35s
uniformly_sampled_quant_bind_relative.config.sjson: 01.87s -> 01.29s
uniformly_sampled_quant_high.config.sjson: 02.27s -> 01.33s
uniformly_sampled_quant_highest.config.sjson: 03.82s -> 01.27s
uniformly_sampled_quant_medium.config.sjson: 01.68s -> 01.27s
uniformly_sampled_quant_mtx_error.config.sjson: 01.68s -> 01.27s
uniformly_sampled_raw.config.sjson: 01.19s -> 01.20s
```

I also tried a more accurate version of propagation which compared every bone with every ancestor.  It resulted in smaller distances, but was more complex(O(N^2)), compression took longer, and memory savings shrank.

I'm confident that introducing error correction within segment compression would improve memory and compression time further.  It seems unlikely that most of the logic in find_optimal_bit_rates would be required anymore.  Perhaps I'll experiment with this later, but the current state of ACL_COMPRESSION_OPTIMIZED is fast enough, and high-quality enough, for our immediate needs.


**ACL_BIT_RATE_EXPANSION**

Expand bit rate options, from [3..19] to [1..23].  

Note that these results assume that ACL_COMPRESSION_OPTIMIZED is enabled.  I don't recommend trying ACL_BIT_RATE_EXPANSION without it.
Additional 2% reduction in memory overall(3% in humans, 2% in creatures).
No changes to _py make.py -regression_test_ timings worth reporting, but some of our creature compression does get slower.  Edge cases are rare, and max out at 7 seconds this time.