@ddeadguyy 2021-06-24
It's worth noting that ACL_BIT_RATE_EXPANSION does slow down unit tests, but ACL_UNIT_TEST_STRICT from https://github.com/nfrechette/acl/pull/369 helps with that.