@nfrechette 2021-06-25
Wow, very impressive savings!
I wonder if this mostly happens because we now modify the raw data to match the constant value which wasn't the case before. This would have left the bit rate optimization trying to reach a target that is unreachable even with full float precision. This would have lead to more higher bit rates being used and more bit rates tested which would explain the memory savings and the compression time speedup. The error seen by the quantization being lower now makes sense since the algorithm has room to actually reach the tweaked raw value. However, it wouldn't lead to a true lower error. The error correction probably helps here.