@nfrechette 2021-06-26
It looks like the debug unit tests have exploded in execution time.
Some runs now take 40mins+ where before everything ran in <10mins.
For example, this run from a few days ago is normal: https://ci.appveyor.com/project/nfrechette/acl/builds/39556495

These are the ones that seem much worse:
31/43 Test #28: unpack_scalarf_uXX_unsafe .................   Passed  496.69 sec
38/43 Test #32: pack_vector4_XX ...........................   Passed  542.31 sec
41/43 Test #42: pack_vector2_XX ...........................   Passed  499.81 sec
43/43 Test #39: pack_vector3_XX ...........................   Passed  815.68 sec