@ddeadguyy 2021-06-26
Yeah, ACL_BIT_RATE_EXPANSION cranks them way up.  It isn't so bad in release(at least not on my machine).  That's what I was referring to in this comment.

> It's worth noting that ACL_BIT_RATE_EXPANSION does slow down unit tests, but ACL_UNIT_TEST_STRICT from #369 helps with that.

ACL_UNIT_TEST_STRICT leverages SIMD to make vector2 ~2x faster, vector3 ~3x faster, and vector4 ~4x faster than scalar.  That could be enough to get debug unit tests under the limit.