@ddeadguyy 2021-06-28
> I could manually bring in the SIMD-leveraging parts though, without making the tests stricter, with comments to replace them with ACL_UNIT_TEST_STRICT in 2.1.

Looks like this plan worked.